+++ src/platform/settings/components/ColorPaletteMessage.vue
