+++ src/platform/settings/constants/coreSettings.ts
@@ -1,9 +1,9 @@
 import { LinkMarkerShape, LiteGraph } from '@/lib/litegraph/src/litegraph'
+import type { SettingParams } from '@/platform/settings/types'
 import type { ColorPalettes } from '@/schemas/colorPaletteSchema'
 import type { Keybinding } from '@/schemas/keyBindingSchema'
 import { NodeBadgeMode } from '@/types/nodeSource'
 import { LinkReleaseTriggerAction } from '@/types/searchBoxTypes'
-import type { SettingParams } from '@/types/settingTypes'
 
 /**
  * Core settings are essential configuration parameters required for ComfyUI's basic functionality.
