+++ src/platform/settings/types.ts
