+++ tests-ui/tests/store/serverConfigStore.test.ts
@@ -2,8 +2,8 @@ import { createPinia, setActivePinia } from 'pinia'
 import { beforeEach, describe, expect, it } from 'vitest'
 
 import { ServerConfig } from '@/constants/serverConfig'
+import type { FormItem } from '@/platform/settings/types'
 import { useServerConfigStore } from '@/stores/serverConfigStore'
-import type { FormItem } from '@/types/settingTypes'
 
 const dummyFormItem: FormItem = {
   name: '',
